refactor: create separate file for createStructureObject function and rename it to createMatrixStructureObject.js, adapt code to use new function.
